Mejor export, demo+cdn, delete+cdn, etc. Iniciando soporte para phpdoc
git-svn-id: http://svn.almidon.org/trunk@356 924a4236-293c-0410-b38f-c0813de1d4fd

+++ php/db.deleterecord.php
@@ -5,15 +5,17 @@
     if (!empty($getfiles)); {
       $tmp = $this->readRecord($id);
       foreach($getfiles as $val) {
-        # CDN? Remove object
-        if (isset($this->dd[$val]['extra']['cdn']) && $this->dd[$val]['extra']['cdn'] === true) {
-          $auth = new CF_Authentication(CDN_USERNAME, CDN_APIKEY);
-          $auth->authenticate();
-          $conn = new CF_Connection($auth);
-          $cloudfiles = $conn->get_container(CDN_REPO);
-          $cloudfiles->delete_object($tmp[$val]);
-        } else {
-          unlink(ROOTDIR . '/files/' . $this->name . '/' . $tmp[$val]);
+        if(!empty($tmp[$val])) {
+          # CDN? Remove object
+          if (isset($this->dd[$val]['extra']['cdn']) && $this->dd[$val]['extra']['cdn'] === true) {
+            $auth = new CF_Authentication(CDN_USERNAME, CDN_APIKEY);
+            $auth->authenticate();
+            $conn = new CF_Connection($auth);
+            $cloudfiles = $conn->get_container(CDN_REPO);
+            $cloudfiles->delete_object($tmp[$val]);
+          } else {
+            unlink(ROOTDIR . '/files/' . $this->name . '/' . $tmp[$val]);
+          }
         }
       }
     }
